// packages/web/source/api.d.ts
export declare interface User {
	id: string
	name: string
	email?: string
	phone?: string
	lastSignedIn: Date
}
// ...
export declare interface Attribute {
	id: string
	name: string
	description: string
	tags: string[]
	conversations: string[]
}
// ...
export declare type BlamedMessage = {
	in: 'question' | 'message'
	id: string
}
// ...
export declare type AttributeSnapshot = {
	value: string | number | boolean
	observer: string | 'bot'
	timestamp: Date
	message?: BlamedMessage
}

/**
 * A class representing an attribute of a certain user.
 *
 * @typedef {object} UserAttribute
 * @property {string} id.required - The attribute ID.
 * @property {string | number | boolean} value.required - The attribute's value.
 * @property {array<AttributeSnapshot>} history - A list of changes that have been made to the attribute's value.
 */
export declare interface UserAttribute {
	id: string
	value: string | number | boolean
	history: AttributeSnapshot[]
}